+++ tests/integration/connect-test.js
@@ -3,6 +3,8 @@ import connect from 'ember-redux/components/connect';
 import hbs from 'htmlbars-inline-precompile';
 import { moduleForComponent, test } from 'ember-qunit';
 
+const { Component } = Ember;
+
 moduleForComponent('count-list', 'integration: connect test', {
   integration: true,
   setup() {
@@ -71,6 +73,29 @@ test('stateToComputed can be used with component level CP if notifyPropertyChang
   assert.equal(this.$('.dyno').text(), 'name: Tom', 'should render new value when local component CP changed and notifyPropertyChange invoked');
 });
 
+test('stateToComputed is not invoked extraneously', function(assert) {
+  let callCount = 0;
+  const stateToComputed = () => {
+    callCount++;
+    return { callCount };
+  }
+  this.register('component:test-component', connect(stateToComputed)(Component.extend({
+    layout: hbs`{{callCount}}`
+  })));
+
+  this.render(hbs`{{test-component attr=attr}}`);
+  assert.equal(this.$().text(), '1');
+  assert.equal(callCount, 1);
+
+  this.set('attr', 'some-change');
+  assert.equal(this.$().text(), '2');
+  assert.equal(callCount, 2);
+
+  this.get('redux').dispatch({ type: 'FAKE-ACTION' });
+  assert.equal(this.$().text(), '3');
+  assert.equal(callCount, 3);
+});
+
 test('the component should truly be extended meaning actions map over as you would expect', function(assert) {
   this.render(hbs`{{count-list}}`);
   let $random = this.$('.random-state');
@@ -84,16 +109,18 @@ test('the component should truly be extended meaning actions map over as you wou
 test('each computed is truly readonly', function(assert) {
   assert.expect(1);
   this.render(hbs`{{count-list}}`);
-  try {
-    this.$('.btn-alter').trigger('click');
-  } catch (e) {
-    assert.ok(e.message.indexOf('Cannot set read-only property') > -1);
-  }
+  Ember.run(() => {
+    assert.throws(() => {
+      this.$('.btn-alter').trigger('click');
+    }, (e) => {
+      return e.message.indexOf('Assertion Failed: Cannot set redux property "low". Try dispatching a redux action instead.') > -1;
+    });
+  });
 });
 
 test('lifecycle hooks are still invoked', function(assert) {
   assert.expect(3);
-  this.register('component:test-component', connect()(Ember.Component.extend({
+  this.register('component:test-component', connect()(Component.extend({
     init() {
       assert.ok(true, 'init is invoked');
       this._super(...arguments);
@@ -119,14 +146,14 @@ test('connecting dispatchToActions only', function(assert) {
   assert.expect(2);
   const dispatchToActions = () => {};
 
-  this.register('component:test-component-1', connect(null, dispatchToActions)(Ember.Component.extend({
+  this.register('component:test-component-1', connect(null, dispatchToActions)(Component.extend({
     init() {
       this._super(...arguments);
       assert.ok(true, 'should be able to connect components passing `null` to stateToComputed');
     }
   })));
 
-  this.register('component:test-component-2', connect(undefined, dispatchToActions)(Ember.Component.extend({
+  this.register('component:test-component-2', connect(undefined, dispatchToActions)(Component.extend({
     init() {
       this._super(...arguments);
       assert.ok(true, 'should be able to connect components passing `undefined` to stateToComputed');
@@ -135,4 +162,61 @@ test('connecting dispatchToActions only', function(assert) {
 
   this.render(hbs`{{test-component-1}}`);
   this.render(hbs`{{test-component-2}}`);
+});
+
+test('connecting dispatchToActions as object should dispatch action', function(assert) {
+  assert.expect(2);
+
+  const dispatchToActions = {
+    up() {
+      assert.ok(true, 'should be able to pass object of functions to dispatchToActions');
+      return {
+        type: 'UP'
+      };
+    },
+    down() {
+      assert.ok(true, 'should be able to pass object of functions to dispatchToActions');
+      return {
+        type: 'DOWN'
+      };
+    }
+  };
+
+  this.register('component:test-dispatch-action-object', connect(undefined, dispatchToActions)(Component.extend({
+    init() {
+      this._super(...arguments);
+    },
+    layout: hbs`
+      <button class="btn-up" onclick={{action "up"}}>up</button>
+      <button class="btn-down" onclick={{action "down"}}>up</button>
+    `
+  })));
+
+  this.render(hbs`{{test-dispatch-action-object}}`);
+  Ember.run(() => {
+    this.$('.btn-up').trigger('click');
+    this.$('.btn-down').trigger('click');
+  });
+});
+
+test('calling deprecated methods shows message', function(assert) {
+  const warnings = [];
+
+  const originalWarn = Ember.warn;
+  Ember.warn = (message, falsy, { id }) => warnings.push({ message, id });
+
+  this.register('component:test-component', connect()(Component.extend({
+    init() {
+      this.handleChange();
+      this.getAttrs();
+      this._super(...arguments);
+    }
+  })));
+
+  this.render(hbs`{{test-component}}`);
+
+  assert.equal(warnings[0].id, 'ember-redux.no-public-handle-change');
+  assert.equal(warnings[1].id, 'ember-redux.no-public-get-attrs');
+
+  Ember.warn = originalWarn;
 });